+++ src/leiningen/viz/cli.clj
@@ -19,6 +19,7 @@
   [["-e" "--selector SELECTOR" "Selector key in :viz map in project.clj" :parse-fn edn/read-string
     :default :default]
    ["-m" "--hide-missing"      "Hide missing nodes in a graph"]
+   ["-o" "--output-file FILE"  "Filename to output image to"]
    ["-s" "--source SOURCE"     "Fully qualified var name (arity-0 fn)"   :parse-fn edn/read-string]
    ["-z" "--zoom-node NODE"    "Sub-graph root node (EDN) to zoom into"  :parse-fn edn/read-string]
    ["-u" "--usage" "Show usage text"]])